HashingException.java: move to model/util, Rename InvalidPasswordException to InvalidLogInException
