+++ src/main/java/seedu/address/logic/parser/PasswordHashUtil.java
@@ -4,7 +4,7 @@ import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 
-import seedu.address.commons.exceptions.HashingException;
+import seedu.address.model.util.HashingException;
 
 /**
  * A utility class for hashing passwords.
